Use correct value for app queue
There are two levels of `get_wait_list_size`, one on the supergroup
level and one on the group level. Through investigation and some
quick load testing it seems that the correct value to use for the
app group's is the deepest level of the `get_wait_list_size` i.e.
on the group level.

My assumption here is that the supergroup's `get_wait_list_size` is
more similar to the top level queue, which rarely fills up.