Prevent index out of range error when passenger process is killed
If any passenger process that is position `n-1` in the list of processes
is removed (usually due to it being killed), the very next call to
`updateProcesses` will panic with an index out of range error.

The two factors at play here are:
* The length of the `found` array.
* The index values of the remaning processes.

If we have 3 processes indexed as follows:
* 0 - `p1`
* 1 - `p2`
* 2 - `p3`

Assume the first process is killed, and we then call `updateProcesses`,
we get the following output:
* 1 - `p2`
* 2 - `p3`

Time has passed and we call `updateProcesses` once again. When
we do so, the `old` process list is as follows:
* 1 - `p2`
* 2 - `p3`

Note that the `old` process list has a length of 2 but contains an
entry with an index that is larger than the length of the map.

When the `found` array is accessed with the `p3` index, we get the
index out of range error.

TLDR;

If you remove `p1` or `p2`, you still have `p3` with an index of 2, yet
the length of the `found` array is 2.